requirements.txt
asgiref==3.2.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-formtools==2.2
django-import-export==2.1.0
django-redis==4.12.1
django-redis-cache==2.1.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
drf-extensions==0.6.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
